Add Components methods is now finished

+++ Meldown Framework/IdFactory.hpp
@@ -33,7 +33,7 @@ namespace Meltdown
 		/// <summary>
 		/// TypeIdFactory generates an unique ID for each type U. Class T represents the domain. 
 		/// Id's generated for types within the same domain are guaranteed to have unique ids.
-		/// For example you could generate type id from components within the domain of BaseComponent. 
+		/// For example you could generate type id from components within the domain of. 
 		/// </summary>
 		template <typename Domain>
 		class TypeIdFactory
@@ -52,7 +52,7 @@ namespace Meltdown
 			static std::size_t GetFlag()
 			{
 				static const std::size_t id = TypeIdFactory<Domain>::template GetId<T>();
-				static const std::size_t flag = exp2(id); //Return id in powers of 2 so that it can represent a flag
+				static const std::size_t flag = static_cast<size_t>(exp2l(static_cast<long double>(id))); //Return id in powers of 2 so that it can represent a flag
 				return flag;
 			};
 		private:
